Rename json-rpc-server to server-process to match class and filename
