+++ extensions/ql-vscode/src/query-server/legacy/queryserver-client.ts
@@ -14,7 +14,7 @@ import {
   WithProgressId,
 } from "../../pure/legacy-messages";
 import { ProgressCallback, ProgressTask } from "../../progress";
-import { ServerProcess } from "../json-rpc-server";
+import { ServerProcess } from "../server-process";
 import { App } from "../../common/app";
 
 type WithProgressReporting = (
